refactor(hermes): split call-parsing primitives into a companion module
hermes-protocol.ts had grown past 3,200 lines. The relaxed-JSON scanning,
repair, key-policy coercion, and boundary-handling primitives move to
hermes-call-parsing.ts, following the existing companion-module pattern
(hermes-argument-schema.ts, hermes-unsafe-pattern.ts). The protocol file
keeps prompt formatting, the streaming state machine, and protocol
assembly. No behavior change.
